jdd_community_backend/requirements-dev.txt
django-debug-toolbar==2.2
django-extensions==3.0.9
